chore: move manage.ts to utils folder
